requirements.txt
azure-ai-projects
azure-identity
python-dotenv
azure-mgmt-logic
fastapi 
pydantic
uvicorn
streamlit